requirements.txt
amadeus==10.1.0
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.8.30
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
exceptiongroup==1.2.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
msgpack==1.1.0
openai==1.46.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
resend==2.4.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
websockets==13.0.1
whitenoise==6.7.0
